Implement Lists
`let a = (1,2,3)` creates a list of 3 elements, and assigns it to `a`.
`list*fn` is the map operator, applies the function to every element in the list, and then returns the new list.
`list/fn` is the filter operator, returning a new list with all elements where the predicate returns true.