Introduce easier list creation
Pushing elements to a list using `while` is very expensive (a list copy for every `list+elem`).

This commit defines the operation `number * func`, returning a new list of length `number`, where every element is the result of the call `function(index)`.

For example, the expr `10 * f(i)i` creates a new list of `(0,1,2,...,8,9)`